1653 - Google Map - rename JS file
