+++ styleguide/source/assets/js/modules/googleMap.js
